feat(frontend): implement interactive stock chart with SMA and RSI indicators
- Add responsive stock price chart with volume visualization
- Include 20-day SMA and 14-day RSI indicators with toggle controls
- Implement time range selection (1M, 3M, 6M, 1Y, 5Y)
- Add interactive tooltips and zoom functionality
- Improve error handling and loading states
- Update UI with Material-UI components and theming

+++ backend/src/main/java/com/tick/backend/repository/StockRepository.java
@@ -4,6 +4,6 @@ import com.tick.backend.model.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {
+public interface StockRepository extends JpaRepository<Stock, Integer> {  // Changed from Long to Integer
     List<Stock> findByTickerOrderByDateAsc(String ticker);
 }
